// src/types/User.ts
export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
}

export interface Company {
  name: string;
}

export interface User {
  id: number;
  username: string;
  name: string;
  email: string;
  phone: string;
  website: string;
  address: Address;
  company: Company;
  liked?: boolean;
  likeCount?: number;
}